Move VariantGeneratorRegistry to own namespace
