+++ src/Message/Handler/ProcessImageHandler.php
@@ -13,9 +13,9 @@ use Setono\SyliusImagePlugin\Exception\ImageProcessingFailedException;
 use Setono\SyliusImagePlugin\File\ImageVariantFile;
 use Setono\SyliusImagePlugin\Message\Command\ProcessImage;
 use Setono\SyliusImagePlugin\Model\ImageInterface;
+use Setono\SyliusImagePlugin\Registry\VariantGeneratorRegistryInterface;
 use Setono\SyliusImagePlugin\Repository\VariantConfigurationRepositoryInterface;
 use Setono\SyliusImagePlugin\VariantGenerator\VariantGeneratorInterface;
-use Setono\SyliusImagePlugin\VariantGenerator\VariantGeneratorRegistryInterface;
 use Setono\SyliusImagePlugin\Workflow\ProcessWorkflow;
 use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
