+++ src/Resources/config/services.xml
@@ -9,6 +9,7 @@
         <import resource="services/event_subscriber.xml"/>
         <import resource="services/message.xml"/>
         <import resource="services/provider.xml"/>
+        <import resource="services/registry.xml" />
         <import resource="services/synchronizer.xml"/>
         <import resource="services/twig.xml"/>
         <import resource="services/variant_generator.xml"/>
