+++ src/Resources/config/services/message.xml
@@ -5,7 +5,7 @@
         <service id="setono_sylius_image.message.handler.process_image"
                  class="Setono\SyliusImagePlugin\Message\Handler\ProcessImageHandler">
             <argument type="service" id="doctrine"/>
-            <argument type="service" id="setono_sylius_image.variant_generator.registry"/>
+            <argument type="service" id="setono_sylius_image.registry.variant_generator"/>
             <argument type="service" id="setono_sylius_image.repository.variant_configuration"/>
             <argument type="service" id="workflow.registry"/>
             <argument type="service">
