+++ src/Resources/config/services/registry.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<container xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://symfony.com/schema/dic/services"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="setono_sylius_image.registry.variant_generator"
+                 class="Setono\SyliusImagePlugin\Registry\VariantGeneratorRegistry">
+        </service>
+    </services>
+</container>
